consolidate API helpers
