+++ src/js/helpers/disasterHelper.js
@@ -12,3 +12,10 @@ export const fetchDEFCodes = () => apiRequest({
 export const fetchOverview = () => apiRequest({
     url: 'v2/disaster/overview/'
 });
+
+export const fetchSpendingOverTime = (params) => apiRequest({
+    isMocked: true,
+    url: 'v2/disaster/spending_over_time/',
+    method: 'post',
+    data: params
+});
